+++ apps/catalog/catalog/src/app/dashboard/dashboard.module.ts
@@ -41,8 +41,7 @@ const routes: Routes = [
       },
       {
         path: 'import', // Import bulk of movies
-        loadChildren: () => import('@blockframes/movie/components/import/import-movie.module')
-          .then(m => m.ImportMovieModule)
+        loadChildren: () => import('@blockframes/import').then(m => m.ImportModule)
       },
       {
         path: 'search',  // Result of a search on the main searchbar
