+++ libs/import/src/lib/components/table-extracted-elements/contracts/contracts.component.scss
