+++ libs/import/src/lib/components/table-extracted-elements/movies/movies.component.scss
