+++ libs/import/src/lib/components/table-extracted-elements/movies/movies.component.ts
@@ -7,7 +7,7 @@ import { Component, Input, ViewChild, OnInit, ChangeDetectionStrategy } from '@a
 import { SelectionModel } from '@angular/cdk/collections';
 import { ViewImportErrorsComponent } from '../view-import-errors/view-import-errors.component';
 import { sortingDataAccessor } from '@blockframes/utils/table';
-import { MovieImportState, SpreadsheetImportError } from '../../import-utils';
+import { MovieImportState, SpreadsheetImportError } from '../../../import-utils';
 import { MovieService } from '@blockframes/movie/+state';
 import { OrganizationQuery } from '@blockframes/organization/+state';
 import { ContractService } from '@blockframes/contract/contract/+state/contract.service';
