+++ libs/import/src/lib/components/table-extracted-elements/organizations/organizations.component.scss
