+++ libs/import/src/lib/components/table-extracted-elements/organizations/organizations.component.ts
@@ -7,7 +7,7 @@ import { Component, Input, ViewChild, OnInit, ChangeDetectionStrategy } from '@a
 import { SelectionModel } from '@angular/cdk/collections';
 import { ViewImportErrorsComponent } from '../view-import-errors/view-import-errors.component';
 import { sortingDataAccessor } from '@blockframes/utils/table';
-import { OrganizationsImportState, SpreadsheetImportError } from '../../import-utils';
+import { OrganizationsImportState, SpreadsheetImportError } from '../../../import-utils';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { AuthService } from '@blockframes/auth/+state';
 import { PublicUser } from '@blockframes/user/+state';
