+++ libs/import/src/lib/components/table-extracted-elements/rights/rights.component.scss
