+++ libs/import/src/lib/components/view-extracted-elements/movies/view-extracted-movies.component.scss
