+++ libs/import/src/lib/components/view-extracted-elements/organizations/view-extracted-organizations.component.scss
