+++ libs/import/src/lib/components/view-extracted-elements/rights/view-extracted-rights.component.html
