+++ tsconfig.json
@@ -56,7 +56,8 @@
       "@blockframes/apps/catalog/*": ["apps/catalog/catalog/src/app/*"],
       "@blockframes/event/*": ["libs/event/src/lib/*"],
       "@blockframes/e2e/*": ["libs/e2e/src/lib/*"],
-      "@blockframes/landing/*": ["libs/landing/src/lib/*"]
+      "@blockframes/landing/*": ["libs/landing/src/lib/*"],
+      "@blockframes/import": ["libs/import/src/index.ts"]
     }
   },
   "angularCompilerOptions": {
